Rename restore_onchain.rs, flatten onchain inner module
